requirements.txt
# Core Django Framework
Django==5.2.6

# Database
mysqlclient==2.2.4

# AI & Machine Learning
google-generativeai==0.8.4
python-dotenv==1.0.1

# Forms & Widgets
django-widget-tweaks==1.5.0

# Image Processing & File Handling
Pillow==10.4.0

# API Development
djangorestframework==3.15.2

# Web Server & Static Files
gunicorn==22.0.0
whitenoise==6.8.1

# E-commerce Specific
django-crispy-forms==2.3
crispy-bootstrap5==0.7

# Utilities
requests==2.32.3
python-decouple==3.8
django-extensions==3.2.3

# CORS & API Security
django-cors-headers==4.3.1

# Filtering & Search
django-filter==24.3

# Authentication & Authorization
django-allauth==65.3.0
django-allauth[socialaccount]==65.3.0

# Caching
django-cacheops==7.1.0
redis==5.0.8

# Background Tasks
celery==5.4.0
django-celery-beat==2.6.0

# Security
django-ratelimit==4.1.0
django-session-timeout==0.1.0

# Testing
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.0

# Development Tools
django-debug-toolbar==4.4.6
django-livereload-server==0.4

# Environment & Configuration
python-dotenv==1.0.1

# Logging
structlog==24.4.0

# Monitoring
sentry-sdk==2.14.0